Added extra functionality on Edit Listing resolved the issue of images on dynamic params
Images - If the url is not valid then the images on header along with the spinner were not working, so to resolve that the / slash is added ahead of  their urls in local.
EditListing - While editing if we wanted to edit just the name of any other field that was not posible due to required field images. Hence the functionality added just to edit any field.